pyproject.toml
[project]
name = "weather-app-mannyyebz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
[tool.hatch.build.targets.wheel]
packages = ["src/weather_app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
